+++ CHANGELOG.md
@@ -30,7 +30,7 @@
 - Introduce a dedicated usage metric for transactions that replaces the duration metric. ([#2571](https://github.com/getsentry/relay/pull/2571), [#2589](https://github.com/getsentry/relay/pull/2589))
 - Restore the profiling killswitch. ([#2573](https://github.com/getsentry/relay/pull/2573))
 - Add `scraping_attempts` field to the event schema. ([#2575](https://github.com/getsentry/relay/pull/2575))
-- Don't aggregate metrics until we received its' project state. ([#2588](https://github.com/getsentry/relay/pull/2588))
+- Don't aggregate metrics until we received its project state. ([#2588](https://github.com/getsentry/relay/pull/2588))
 
 ## 23.9.1